+++ src/main/java/com/thewhite/utilitystorage/action/rating/add/AddRatingActionArgument.java
@@ -1,10 +1,10 @@
 package com.thewhite.utilitystorage.action.rating.add;
 
 import com.thewhite.utilitystorage.model.rating.NumberPoints;
-import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Value;
 
+import javax.validation.constraints.NotNull;
 import java.util.UUID;
 
 @Value
