+++ src/main/java/com/thewhite/utilitystorage/api/rating/dto/AddRatingDto.java
@@ -2,10 +2,10 @@ package com.thewhite.utilitystorage.api.rating.dto;
 
 import com.thewhite.utilitystorage.model.rating.NumberPoints;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Value;
 
+import javax.validation.constraints.NotNull;
 import java.util.UUID;
