+++ src/main/java/com/thewhite/utilitystorage/api/utilitystorage/dto/UpdateUtilityDto.java
@@ -3,9 +3,8 @@ package com.thewhite.utilitystorage.api.utilitystorage.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Value;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
 
+import javax.validation.constraints.NotNull;
 import java.util.Set;
 import java.util.UUID;
