+++ src/main/java/com/thewhite/utilitystorage/repository/UtilityStorageRepository.java
@@ -4,10 +4,9 @@ import com.thewhite.utilitystorage.model.utilityStorage.UtilityStorage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
-@Component
 public interface UtilityStorageRepository extends JpaRepository<UtilityStorage, UUID>, QuerydslPredicateExecutor<UtilityStorage>, PagingAndSortingRepository<UtilityStorage, UUID> {
 }
