+++ src/main/resources/application.properties
@@ -4,3 +4,5 @@ spring.datasource.password=1234
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
+
+logging.level.com.example=DEBUG
