+++ src/test/java/com/thewhite/utilitystorage/service/rating/RatingServiceIT.java
@@ -2,7 +2,6 @@ package com.thewhite.utilitystorage.service.rating;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.spring.api.DBRider;
 import com.jupiter.tools.spring.test.postgres.annotation.meta.EnablePostgresIntegrationTest;
 import com.thewhite.utilitystorage.model.rating.NumberPoints;
 import com.thewhite.utilitystorage.model.rating.Rating;
@@ -12,21 +11,17 @@ import org.assertj.core.api.SoftAssertions;
 import org.assertj.core.api.junit.jupiter.SoftAssertionsExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 import java.util.UUID;
 
-@SpringBootTest
 @EnablePostgresIntegrationTest
 @ExtendWith(SoftAssertionsExtension.class)
-@DBRider
+
 class RatingServiceIT {
 
-    @Autowired
     RatingService service;
 
     @Test
